Move template rendering code into a submodule of `tmt.utils`
